Removed the basic_request class
