+++ include/siddiqsoft/private/rest_request.hpp
@@ -51,20 +51,156 @@
 #include "siddiqsoft/date-utils.hpp"
 
 #include "restcl_definitions.hpp"
-#include "basic_request.hpp"
 
 namespace siddiqsoft
 {
     /// @brief A REST request utility class. Models the request a JSON document with `request`, `headers` and `content` elements.
-    class rest_request : public basic_request
+    /// Essentially we're a convenience wrapper on the rest_request.
+    class rest_request
     {
+    protected:
+        /// @brief Not directly constructible; use the derived classes to build the request
+        rest_request()
+            : nlohmann::json({{"request", {{"method", nullptr}, {"uri", nullptr}, {"protocol", nullptr}}},
+                              {"headers", nullptr},
+                              {"content", nullptr}})
+        {
+        }
+
+        /// @brief Constructs a request with endpoint string into internal Uri object
+        /// @param s Valid endpoint string
+        explicit rest_request(const std::string& s) noexcept(false)
+            : nlohmann::json({{"request", {{"method", nullptr}, {"uri", nullptr}, {"protocol", nullptr}}},
+                              {"headers", nullptr},
+                              {"content", nullptr}})
+            , uri(s) { };
+
+        /// @brief Not directly constructible; use the derived classes to build the request
+        /// @param s The source Uri
+        explicit rest_request(const Uri<char>& s) noexcept(false)
+            : nlohmann::json({{"request", {{"method", nullptr}, {"uri", nullptr}, {"protocol", nullptr}}},
+                              {"headers", nullptr},
+                              {"content", nullptr}})
+            , uri(s) { };
+
+    public:
+        /// @brief Set the content (non-JSON)
+        /// @param ctype Content-Type
+        /// @param c The content
+        /// @return Self
+        rest_request& setContent(const std::string& ctype, const std::string& c)
+        {
+            if (!c.empty()) {
+                this->at("headers")[HF_CONTENT_TYPE]   = ctype;
+                this->at("headers")[HF_CONTENT_LENGTH] = c.length();
+                this->at("content")                    = c;
+            }
+            return *this;
+        }
+
+        /// @brief Set the content to json
+        /// @param c JSON content
+        /// @return Self
+        rest_request& setContent(const nlohmann::json& c)
+        {
+            if (!c.is_null()) {
+                this->at("content")                    = c;
+                this->at("headers")[HF_CONTENT_LENGTH] = c.dump().length();
+                // Make sure we do not override existing value
+                if (!this->at("headers").contains(HF_CONTENT_TYPE)) this->at("headers")[HF_CONTENT_TYPE] = "application/json";
+            }
+            return *this;
+        }
+
+
+        std::string getContent() const
+        {
+            // Build the content to ensure we have the content-type
+            if (this->at("content").is_object()) {
+                return this->at("content").dump();
+            }
+            else if (this->at("content").is_string()) {
+                return this->at("content").get<std::string>();
+            }
+
+            return {};
+        }
+
+
+        /// @brief Encode the headers to the given argument
+        /// @param rs String where the headers is "written-to".
+        void encodeHeaders_to(std::string& rs) const
+        {
+            for (auto& [k, v] : this->at("headers").items()) {
+                if (v.is_string()) {
+                    std::format_to(std::back_inserter(rs), "{}: {}\r\n", k, v.get<std::string>());
+                }
+                else if (v.is_number_unsigned()) {
+                    std::format_to(std::back_inserter(rs), "{}: {}\r\n", k, v.get<uint64_t>());
+                }
+                else if (v.is_number_integer()) {
+                    std::format_to(std::back_inserter(rs), "{}: {}\r\n", k, v.get<int>());
+                }
+                else {
+                    std::format_to(std::back_inserter(rs), "{}: {}\r\n", k, v.dump());
+                }
+            }
+
+            std::format_to(std::back_inserter(rs), "\r\n");
+        }
+
+
+        /// @brief Encode the request to a byte stream ready to transfer to the remote server.
+        /// @return String
+        std::string encode() const
+        {
+            std::string rs;
+            std::string body;
+            auto&       hs = this->at("headers");
+            auto&       rl = this->at("request");
+
+            // Request Line
+            std::format_to(std::back_inserter(rs),
+                           "{} {} {}\r\n",
+                           rl["method"].get<std::string>(),
+                           rl["uri"].get<std::string>(),
+                           rl["version"].get<std::string>());
+
+            // Build the content to ensure we have the content-type
+            if (this->contains("content")) {
+                if (this->at("content").is_object()) {
+                    body = this->at("content").dump();
+                }
+                else if (this->at("content").is_string()) {
+                    body = this->at("content").get<std::string>();
+                }
+            }
+
+            // Headers..
+            encodeHeaders_to(rs);
+
+            // Finally the content..
+            if (!body.empty()) {
+                std::format_to(std::back_inserter(rs), "{}", body);
+            }
+
+            return rs;
+        }
+
+
+    public:
+        friend std::ostream& operator<<(std::ostream&, const rest_request&);
+
+    public:
+        Uri<char, AuthorityHttp<char>> uri;
+
     public:
         /// @brief Constructor with only endpoint as string argument
         /// @param endpoint Valid endpoint string
         rest_request(const std::string& verb,
                      const std::string& endpoint,
                      const std::string& proto = HTTP_PROTOCOL_VERSIONS[1]) noexcept(false)
-            : basic_request(endpoint)
+            : uri(endpoint)
         {
             // Build the request line data
             this->at("request") = {{"uri", uri.urlPart}, {"method", verb}, {"protocol", proto}};
@@ -82,7 +218,7 @@ namespace siddiqsoft
         /// @param endpoint Fully qualified http schema uri
         /// @param h Optional json containing the headers
         rest_request(const std::string& verb, const Uri<char>& endpoint, const nlohmann::json& h = nullptr) noexcept(false)
-            : basic_request(endpoint)
+            : uri(endpoint)
         {
             // Set the headers if present (we add some defaults below if not provided or missing param)
             if (h.is_object() && !h.is_null()) {
@@ -118,23 +254,26 @@ namespace siddiqsoft
         explicit rest_request(const Uri<char>& endpoint, const nlohmann::json& h, const std::string& c) noexcept(false)
             : rest_request(endpoint, h)
         {
-            if (h.is_null() || (h.is_object() && !h.contains("Content-Type"))) this->at("headers")["Content-Type"] = "text/plain";
-            this->at("headers")[HF_CONTENT_LENGTH] = c.length();
-            this->at("content")                    = c;
+            if (!c.empty()) {
+                if (h.is_null() || (h.is_object() && !h.contains("Content-Type")))
+                    this->at("headers")["Content-Type"] = "text/plain";
+                this->at("headers")[HF_CONTENT_LENGTH] = c.length();
+                this->at("content")                    = c;
+            }
         }
 
         /// @brief Constructor with arbitrary content. If the header Content-Type is missing then we use text/plain
         /// @param endpoint Fully qualified http schema uri
         /// @param h Required json containing the headers. At least Content-Type should be set.
         /// @param c Required string containing the content. Make sure to set "Content-Type"
-        explicit rest_request(const Uri<char>& endpoint, const nlohmann::json& h, const char* c) noexcept(false)
+        explicit rest_request(const Uri<char>& endpoint, const nlohmann::json& h, const char* srcContent) noexcept(false)
             : rest_request(endpoint, h)
         {
-            if (c != nullptr) {
+            if (srcContent != nullptr) {
                 if (h.is_null() || (h.is_object() && !h.contains("Content-Type")))
                     this->at("headers")["Content-Type"] = "text/plain";
-                this->at("headers")[HF_CONTENT_LENGTH] = strlen(c);
-                this->at("content")                    = c;
+                this->at("headers")[HF_CONTENT_LENGTH] = strlen(srcContent);
+                this->at("content")                    = srcContent;
             }
             else {
                 this->at("headers")[HF_CONTENT_LENGTH] = 0;
@@ -142,6 +281,11 @@ namespace siddiqsoft
         }
     };
 
+    inline std::ostream& operator<<(std::ostream& os, const rest_request& src)
+    {
+        os << src.encode();
+        return os;
+    }
 
     namespace restcl_literals
     {
