+++ include/siddiqsoft/private/restcl_unix.hpp
@@ -34,7 +34,6 @@
 #include "nlohmann/json.hpp"
 
 #include "restcl_definitions.hpp"
-#include "basic_request.hpp"
 #include "basic_response.hpp"
 #include "basic_restclient.hpp"
 #include "rest_request.hpp"
@@ -132,7 +131,7 @@ namespace siddiqsoft
         /// @brief Implements an asynchronous invocation of the send() method
         /// @param req Request object
         /// @param callback The method will be async and there will not be a response object returned
-        void send(basic_request&& req, std::optional<basic_callbacktype> callback = std::nullopt) override
+        void send(rest_request&& req, std::optional<basic_callbacktype> callback = std::nullopt) override
         {
             if (!isInitialized) throw std::runtime_error("Initialization failed/incomplete!");
 
@@ -147,7 +146,7 @@ namespace siddiqsoft
         /// @brief Implements a synchronous send of the request.
         /// @param req Request object
         /// @return Response object only if the callback is not provided to emulate synchronous invocation
-        [[nodiscard]] basic_response send(const basic_request& req) override
+        [[nodiscard]] basic_response send(const rest_request& req) override
         {
             auto rc {0};
